% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_grid_mapping.R
\name{AgeGridMapping}
\alias{AgeGridMapping}
\title{Build an exact age-grid mapping}
\usage{
AgeGridMapping(
  from_age_groups,
  to_age_groups,
  open_ended = c("include", "error")
)
}
\arguments{
\item{from_age_groups}{Source age structure.}

\item{to_age_groups}{Target age structure.}

\item{open_ended}{How to handle open-ended age groups.}
}
\value{
An \code{AgeGridMapping} object.
}
\description{
Build an exact age-grid mapping
}
