% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_grid_mapping.R
\name{aggregate_cumulative_flows_age_grid}
\alias{aggregate_cumulative_flows_age_grid}
\title{Aggregate cumulative-flow output to a reporting age grid}
\usage{
aggregate_cumulative_flows_age_grid(x, mapping)
}
\arguments{
\item{x}{Cumulative-flow data frame.}

\item{mapping}{An \code{AgeGridMapping} object.}
}
\value{
Aggregated cumulative-flow data frame.
}
\description{
Aggregate cumulative-flow output to a reporting age grid
}
