% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_grid_mapping.R
\name{aggregate_demography_trajectory_age_grid}
\alias{aggregate_demography_trajectory_age_grid}
\title{Aggregate a demography trajectory to a reporting age grid}
\usage{
aggregate_demography_trajectory_age_grid(x, mapping)
}
\arguments{
\item{x}{Data frame with \code{time}, \code{age_group}, and \code{population}.}

\item{mapping}{An \code{AgeGridMapping} object.}
}
\value{
Aggregated demography trajectory.
}
\description{
Aggregate a demography trajectory to a reporting age grid
}
