% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_grid_mapping.R
\name{aggregate_epidemic_trajectory_age_grid}
\alias{aggregate_epidemic_trajectory_age_grid}
\title{Aggregate an epidemic trajectory to a reporting age grid}
\usage{
aggregate_epidemic_trajectory_age_grid(x, mapping)
}
\arguments{
\item{x}{Data frame with \code{time}, \code{compartment}, \code{age_group}, and \code{value}.}

\item{mapping}{An \code{AgeGridMapping} object.}
}
\value{
Aggregated epidemic trajectory.
}
\description{
Aggregate an epidemic trajectory to a reporting age grid
}
