% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_grid_mapping.R
\name{aggregate_population_summary_age_grid}
\alias{aggregate_population_summary_age_grid}
\title{Aggregate an age-group population summary to a reporting age grid}
\usage{
aggregate_population_summary_age_grid(x, mapping)
}
\arguments{
\item{x}{Data frame with \code{time}, \code{age_group}, and \code{value}.}

\item{mapping}{An \code{AgeGridMapping} object.}
}
\value{
Aggregated population summary.
}
\description{
Aggregate an age-group population summary to a reporting age grid
}
