% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{demography_population_vector}
\alias{demography_population_vector}
\title{Get an age-specific population vector}
\usage{
demography_population_vector(demography, time)
}
\arguments{
\item{demography}{An \code{agepi_demography} object.}

\item{time}{Exact time point to retrieve.}
}
\value{
Numeric vector of population values.
}
\description{
Returns population values for one exact time point. Values are ordered
and named by the demography object's age groups.
}
