% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{demography_times}
\alias{demography_times}
\title{Get demography time points}
\usage{
demography_times(demography)
}
\arguments{
\item{demography}{An \code{agepi_demography} object.}
}
\value{
Numeric vector of available time points.
}
\description{
Get demography time points
}
