% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_process.R
\name{validate_demographic_process}
\alias{validate_demographic_process}
\title{Validate a demographic process configuration}
\usage{
validate_demographic_process(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
The input invisibly if validation succeeds.
}
\description{
Validate a demographic process configuration
}
