% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_rates.R
\name{validate_fertility_schedule}
\alias{validate_fertility_schedule}
\title{Validate a fertility schedule}
\usage{
validate_fertility_schedule(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
The input invisibly if validation succeeds.
}
\description{
Validate a fertility schedule
}
