% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_rates.R
\name{validate_migration_schedule}
\alias{validate_migration_schedule}
\title{Validate a migration schedule}
\usage{
validate_migration_schedule(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
The input invisibly if validation succeeds.
}
\description{
Validate a migration schedule
}
