% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_rates.R
\name{validate_mortality_schedule}
\alias{validate_mortality_schedule}
\title{Validate a mortality schedule}
\usage{
validate_mortality_schedule(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
The input invisibly if validation succeeds.
}
\description{
Validate a mortality schedule
}
