% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_demographic_inputs.R
\name{wpp_dataset}
\alias{wpp_dataset}
\title{Load a dataset from the optional wpp2024 package}
\usage{
wpp_dataset(dataset)
}
\arguments{
\item{dataset}{Single dataset name to load from \code{wpp2024}.}
}
\value{
A data frame.
}
\description{
Load a dataset from the optional wpp2024 package
}
